requirements.txt
# Install this first
git+https://github.com/InjectiveLabs/sdk-python.git@v1.9.0#egg=injective-py
python-telegram-bot==20.7
Flask==3.0.0
Flask-Cors==4.0.0
aiohttp>=3.9.4
APScheduler==3.6.3
urllib3>=2.0.0
pyngrok>=7.2.3
bech32==1.2.0
python-dotenv==1.0.0